test: consolidate duplicate and shared-setup tests
Reduce cargo-test compile/link cost by merging test binaries and
sharing expensive setup, and drop genuinely redundant unit tests.

- pyre-jit: merge 6 *_gc_stress.rs into one gc_stress.rs behind a
  shared run_harness; each test still runs on its own 256MiB worker
  thread, preserving per-test isolation.
- majit-macros: merge driver/jit_module/jit_struct test binaries into
  a single mod-wrapped macros.rs.
- majit-translate: hoist src-tree/corpus into OnceLocks and share a
  compile helper and reachable_blocks via tests/common.
- majit-metainterp: extract build_dispatch_minimal and
  build_oparg_minimal fixtures in jit_interp_dispatch_ir_shape.rs.
- pyre-interpreter/pyre-object: delete redundant unit tests; merge the
  setattr get/overwrite and extended-arg decode cases.
